Not all messages have labels (apparently)
So deal with it gracefully.

Also enabled tqdm bar.  Not working right ATM since I don't know a
priori how many messages there are.  I should run a test to see how long
it takes to list all the messages to get an accurate count but it seems
like a useless waste of bandwidth, CPU, and time.