Implemented a faster way to update labels
Once the user has synch-ed the metadata once, the bulk of a full sync is
to compare the labels and update them as necessary.  This is mostly how
fast we can read the labels from the database.

I implemented a new mechanism for doing so that only reads/updates the
associations table.  This required using SQL syntax in sqlalchemy--or at
least I could not figure out a way to do so with ORM.

I am leaving for now a third variation that is a mix of the two
approaches (although it doesn't quite work right).  Leaving there so it
get checked into git and will remove it in a subsequent checkin.