Added code for incremental updates
First fetch all new changes, then merge the updates, and then apply the
updates.  This means that if a message gets added and deleted in the
same history sequence, then the message never gets added to the local
database.  Similarly adding a removing a label (both of which happen
within this incremental update) will never show up in the local
database.