Insert messages in chronological order
gnus wants article numbers, for which I plan to use the row ID, to be
sorted by time (increasing with time).  This requires that full
updates insert articles in chronological order.

Unfortunately the gmail API provides them in descending order (most
recent messages first).  Which makes sense for most clients, but not
here.

So, fetch all the ID's, sort them (the message ID's seem to be
increasing) and then create the DB rows.