First cut implementation os raw data caching
Added a hidden property to message (_raw) which holds a compressed
version of the message contents (raw message so it includes all headers
and body).  This property is marked a deferred so it won't get read in
on a lookup unless explicitly asked for.  To avoid double queries I
added a new argument to find() so it can fetch raw on lookup when
needed.