Revamp email parsing
My first attempt was broken because I split the header on comas which is
a no-no since the comma can be inside a quoted string.

So rely instead on getaddresses() from the email package.

Also cleaned up the handling of headers which have slightly different
spelling (mostly capitalized vs not).