Function to create all rows in the Message table
I don't have a way to handle failed batch_executor() requests that fail
because messages need to be created in order (so id's are monotonically
increasing in time as gnus wants).  I have bee exploring the idea of
creating all the entries in the Message table first (using sorted Google
id's) and then updating those rows when we get the data.

This change adds a function to create all the rows in the Message table
given a (sorted) list of Google id's.