Restructured the code to look like a python package
Code for mail sync is now in nnsync.  The web app is in nngmail (which I
should perhaps rename to web or app).

Moved most of the code for nnserver to a new module (server).  This
should probably now go by the wayside since I plan to use the flask app.