Add an include argument to serialize
In some cases it is easier to specify a white list rather than a
blacklist of attributes.

This along with the sub-classed JSON encoder makes json responses easy.

Added a few more constraints to the models so things look nice.  For
Addressees, skip the addressee and only serialize the contact.  The type
of addressee should be obvious from the enclosing envelope.