Don't create a closure on self when starting sync
I don't know if this makes a difference or not, but don't create a
closure on self (the object on which sync got called) since we don't
need to keep that specific instance.  Instead just create a new object
in the new thread.

I did this to avoid confusion on what thread is accessing which
objects.