Add a retry mechanism for nngmail-request-article
Since messages (articles) may not be locally available, need a mechanism
to either retry or wait for it to be available.

This change adds a retry mechanism.  But it is awkward because the time
to get the message will vary.  I had to add a sleep-for in the middle of
the loop to get stuff to work with small messages.  I suspect ti would
be better to synchronously wait until while the message is fetched.