Ignore the ID's GNus gives us and fetch that number of articles
Gnus gets hopelessly confused by the gaps in ID's--which happen because
all messages of all accounts are in the same table.  So rather than
asking for specific ID's, we ask for that number of messages.

This still doesn't work because Gnus ignores the messages it receives ad
only displays the ones it thinks it should display.  Argh!