Add routes with names instead of ID's for labels
The elisp code has a bunch of mapping from name (or nickname) to an id.
Which can probably be done more easily in python/SQL.  As a first step,
add routes that take as input a label name (and account id or nickname).

This is just dipping my toes in the water.