Add support for reading messages via helm
helm-nngmail-unread queries the local server for all unread messages in
all accounts and provides a helm buffer for selecting one of the
messages.  Currently the actions don't do anything but I plan to hook
them up to gnus to be able to read, mark as read, or delete the message.