Add support for prompting the user for a label to process
By default look for messages with the UNREAD label.  With prefix arg,
prompt the user for which label to process.

Moved the nngmail-specific functions to nngmail.el

Use the message alist as the candidate rather than constructing a new
list.  But add a server and group item to the alist since we probably
will need them in the action.