Convert message table to a composite primary key
When synch-ing multiple accounts I was ending up with big gaps in
article numbers (or primary id values in the DB) because the id was
unique across all accounts.

This change converts the table to a composite primary key (id,
account_id).  This allows duplicate keys across accounts and avoids
those big gaps in the article number space.