Implemented message update and delete
This change builds up support for updating and deleting messages.
However, the requests will always fail because they have not yet been
implemented in remote.py (for now there are three stubs that always
return an error).

In a separate branch I will add support for making write requests from
Gmail.  This makes it possible to propagate changes up to the call to
Gmail.

message.py first tries to update the remote and then does the local
updates.  Since the Gmail request fails, it never attempts to update the
local database.