Add a mechanism for adding (injecting) properties during serialization
I haven't found a nice clean way to add URL's to the serializaed JSOn
representation of a database object.  Mostly because I don't want to
have url_for() in the model definition (that seems to break the
separation between mode/controller).

This changes adds a mechanism for injecting (or adding) extra properties
to a serialized object.  The value of the property is evaluated
dynamically so we can extract dynamic properties.  In particular, it is
possible to call url_for() to get the URL for an object.,