Call nnoo-change-server open opening a server
And call nnoo-close-server when closing a server (which now happens if
we don't find the account on open).

Calling nnoo-change-server I think is supposed to set the variables
defined in the secondary method definition.  And it indeed does this,
but only in the context of the function calling nnoo-change-server.  So
perhaps I need to do this for all functions that make requests?

Changed the name of the variables so they better match nnimap (and what
I had in my _gnus file).

Made nngmail-base-url a function so it can be called dynamically (in
case the address or port parameters change).  Though this is somewhat at
odds with the proxy server always being local.