Remove retry of message get
Since message get is now synchronous, there's no need for a retry.  It
is the responsibility's of the server to give us a message.

Along with this change, skip trying to fetch messages if Gmail is
unavailable (likely because we are offline).  Return an indication the
message is currently unavailable.