Store the base-url for a server (account) in server-open
The user can override the -address and -port parameters in
gnus-secondary-select-methods so we need to remember those parameters.

There is one hole, however.  If something fetches a resource using an
account ID (numeric) then it will get the default base URL (for now).  I
need to add a mechanism for getting the account nickname from the id.