Make link from Addressee table to message use composite primary key
When I moved to composite primary keys (so each account has nice
not-too-gappy article numbers) I overlooked that the Addressee table
needed to use the composite key for the join.  Otherwise it is ambiguous
which message it refers to.

This change adds a new column (account_id) to Addressee and specifies
primaryJoin constraints to make everything work.