Converted from composite primary key to an explicit article ID
I switched to using composite keys for messages (id, account_id).  But
this was hard to work with and hard to describe in the code (I couldn't
figure out how to describe the label_association and Addressee
relationships in SQLAlchemy).

Instead use an explicit article ID (article_id) which is per account.
The Message.id column is now unique across all accounts and can be used
for relationships.

Almost everything is working except for a couple of nngmail tests.
Checking stuff in as a semi-working baseline before looking at these
tricky failing tests.