Performance optimization and functional bug fixes for flags
Only look for UNREAD messages with the label specified in the
parameters.

Fetching flags for my personal account (~50k messages) was slow.  I
thought this was due to sqlite but turned out to be mostly python.

I think the biggest gain was I got rid of unseen (which Gnus seems to
ignore anyway) since it requires mucking around with very large sets.
Instead query the database for the (min, max) article numbers and fill
in the rest.

i also optimized the other queries a little.