Only select unread messages as candidates
Using all UNREAD messages as candidates doesn't work well because there
can be thousands of unread messages in TRASH and SPAM.  So instead look
for UNREAD messages in INBOX (by default, can be changed to something
else).

Which brings up that I should probably have a way to do local queries so
I can specify a list of present or absent tags.