Add support for incremental updates of unexist
Querying the database for all existing ID's for a group (or label) is
expensive.  For INBOX on my personal account it takes 2+ seconds.
Fortunately we don't need to do this all the time.

For now I am trying incremental updates.  This works for new messages
(new article ID's created after the timestamp).  But it does not handle
older messages that get deleted or moved out of the group.  I'm not sure
how that will work out but committing this for now since it makes
selecting INBOX feel fast (without it there's a significant pause before
entering the group in Gnus).