Remove automatic timestamp insert on update
I use the updated field to determine whether to cache messages or not.
Which requires I be able to not do the update (e.g. when removing the
article body on cache eviction).

I could not get SA to skip the timestamp insertion.  The docs say it
should skip it if the field with the onupdate trigger is present in the
SET to update.  But it didn't seem to work.

Instead, manually update the timestamp when changing labels or caching
the message.