Sort messages in create to ensure they are in-order
Gnus requires article ID be nicely behaved.  So let's move the sorting
to the code that inserts stuff into the DB.

Also don't depend on the objects returned by the DB to be in the same
order as the messages e get from Gmail.  In this case sort the ID's as
strings (it's easier).  It doesn't matter what the order is so long as
both lists are in the same order.