New implementation of incremental mark synchronization
Rather than using a date, pass the max article ID from the previous time
we asked for marks.  This is still not right because older articles may
have been deleted.  I plan to fix this in a subsequent change (always
return a minimum number of articles).

The marks method now returns stuff with a bit of structure since it also
returns the lowest numbered article ID for which marks are included and
the current active range (which we have for free anyway).