Reset the http object when a connection fails
httplib2 keeps sockets hanging around.  When a connection fails this
causes subsequent calls to fail as well, even if the issue has gone
away.  This is most obvious when trying to reconnect after suspend.  The
first connection attempt fails (probably because the networking layer is
not yet fully up) and the next two or three connections fail as well.