Add a messages-url property to the accounts JSON representation
I want to add support for creating "newsgroups" based on a SQL query.
The most basic group would be one that looks for messages with label
(like I have now).  But this would open up the opportunity for more
complex groups.

To do this I want to remove the dependence on label.  Since Gnus asks
for specific headers (it passes a list of article ID's) I can use the
more generic URL endpoint.