Speed up message retrieval
helm-nngmail is slow when reading labels that have lots of messages.  So
optimize this a bit by not loading attributes which I don't use in the
client (to_, cc, bcc, labels).  These attributes require reading
additional tables and it makes the query slow.

In my testing a query with joinedload for labels was .45s.  A query that
skips labels is .21s.

I kept joinedload for sender which is used by the client.  Using
joinedload means we only got to the database once.

Changed the name of the URL attributes in the info serialization.