Generalize check for start_aid < min queried article ID
I used to special case start_aid being 1 (request all marks).  But this
is a more general case of the passed start_aid falling in a unexist
hole.

For a full fetch, in general the start_aid is 1 but the first valid
article ID > 1.  Need to make sure 1:first_aid is marked as unexist.
The same applies anytime start_aid fall into a hole.