Cleanup relationships and deletes further
Move all relationship declarations to the parent (the entity not having
a foreign key column) and setup passive cascaded deletes appropriately.
Deleting an account is now instantaneous.

This also cleans up thread_id not being set when creating messages.  I
want to make sure the columns are never null so I added a not nullable
constraint to the table.  That required I add a dummy account and a
dummy thread when creating the database.