Increase sqlite3 timeout
Periodically one thread will raise an exception because the DB is
locked.  For now apply a band-aid and increase the timeout (during which
time the DB client will wait for the lock).  If the problem persists, I
can look at better solutions.

I believe this is caused by contention between the sync threads and the
client threads.