Initial implementation of back-off on user rate errors
Google limits the request rate for Gmail.  If we read messages too
quickly it returns an error and the client needs to back-off and retry
later.

I had never run into these errors before, but they seem to be popping up
frequently now.

This change is a first (and flawed) attempt to deal with the problem.
On a user rate error the code will wait X seconds and try again.  It
will continue to double the delay between requests until it reaches a
maximum or succeeds.