Add native JSON support
Use native JSON parsing (when available) to speedup communication with
the server.  Most often parsing shouldn't be a big burden, but there
are a few instances when messages get lengthy.

The code only does PUT requests for updating "marks" so switching to
native serialization is not a big win.  But let's clean it up anyway.