Add support for restarting import or full synch
Full synchs of large mailboxes take some time (20 minutes for mine).
This change adds support for restarting the operation from wherever we
left off.  It works by processing the gids in ascending order and
storing in the database the last gid that was processed.

This only works if the history ID doesn't change between the interrupt
and the restart but I have an idea for how to improve things.