Convert google_id to an int for storing in the database
I used to keep the google_id of a message as a string since that is how
the API makes them available.  But they seem to represent an integer.
Convert the number from a hex string to an int when writing to the
database and undo the conversion when reading.  By storing the value as
an integer, I can compare them and sort them.