Allow partial pulls to re-start is history has moved forward
My first implementation of partial pulls (a pull that restarts a failed
r interrupted full pull) could only proceed if history (according to
Google) hadn't moved forward.  Which doesn't work so well for accounts
that receive a lot of email.

This implementation of partial pulls adds support for completing an
interrupted full pull and then finish with an incremental pull.  This
means we can almost always restart an interrupted full pull.